+++ example/moduleReUse/ActionTypes.js
@@ -1,7 +1,0 @@
-const ADD_PANEL = "ADD_PANEL";
-const DEL_PANEL = "DEL_PANEL";
-
-export default {
-  ADD_PANEL,
-  DEL_PANEL
-};