+++ example/moduleReUse/ActionTypes.ts
@@ -1,0 +1,6 @@
+enum ActionTypes {
+  ADD_PANEL = "ADD_PANEL",
+  DEL_PANEL = "DEL_PANEL"
+}
+
+export default ActionTypes;