+++ example/moduleReUse/actions.ts
@@ -1,13 +1,15 @@
 import ActionTypes from "./ActionTypes";
 
-export function addPanel() {
+function addPanel() {
   return {
     type: ActionTypes.ADD_PANEL
   };
 }
 
-export function delPanel() {
+function delPanel() {
   return {
     type: ActionTypes.DEL_PANEL
   };
 }
+
+export default { addPanel, delPanel };