+++ example/moduleReUse/opPanel/actions.js
@@ -1,6 +1,0 @@
-export function addCnt(num) {
-  return {
-    type: "ADD_CNT",
-    num
-  };
-}