+++ example/moduleReUse/opPanel/actions.ts
@@ -1,0 +1,8 @@
+function addCnt(num) {
+  return {
+    type: "ADD_CNT",
+    num
+  };
+}
+
+export default { addCnt };