+++ example/moduleReUse/opPanel/index.ts
@@ -1,5 +1,5 @@
 import { bundleToComponent } from "redux-arena/tools";
-import * as actions from "./actions";
+import actions from "./actions";
 import OpPanel from "./OpPanel";
 
 export default bundleToComponent({