+++ example/passDownStateAndActions/actions.ts
@@ -1,12 +1,12 @@
 import ActionTypes from "redux-arena/ActionTypes";
 
-export function addCnt() {
+function addCnt() {
   return {
     type: "ADD_CNT"
   };
 }
 
-export function clearCnt() {
+function clearCnt() {
   return {
     type: ActionTypes.ARENA_SCENE_SET_STATE,
     state: {
@@ -14,3 +14,5 @@ export function clearCnt() {
     }
   };
 }
+
+export default { addCnt, clearCnt };