+++ src/core/enhancedRedux/createPropsPicker.ts
@@ -1,9 +1,13 @@
 import { ActionCreatorsMapObject } from "redux";
-import { PropsPicker, ConnectedAction } from "../types";
-import { CurtainReduxInfo, CurtainMutableObj } from "../reducers/types";
+import { PropsPicker, ConnectedAction, StateDict, ActionsDict } from "../types";
+import {
+  CurtainReduxInfo,
+  CurtainMutableObj,
+  RootState
+} from "../reducers/types";
 
 function defaultPropsPicker<S>(
-  { _arenaScene: state }: { _arenaScene: S },
+  { _arenaScene: state }: StateDict<S>,
   { _arenaScene: actions }: { _arenaScene: {} }
 ): DefaultPickedProps<S> {
   return Object.assign({}, state, {
@@ -11,12 +15,31 @@ function defaultPropsPicker<S>(
   });
 }
 
+function getRelativeLevel(name: string) {
+  let result = name.match(/^\$(\d+)$/);
+  return result && parseInt(result[1]);
+}
+
+function getLevelState(
+  rootState: RootState,
+  reducerKey: string,
+  levelNum: number
+) {
+  if (levelNum === 0) return rootState[reducerKey];
+  let { stateTree, stateTreeDict } = rootState.arena;
+  let path = stateTreeDict
+    .getIn([reducerKey, "path"])
+    .skipLast(2 * levelNum)
+    .toList();
+  return stateTree.getIn(path);
+}
+
 export type DefaultPickedProps<S> = {
   actions: Record<string, ConnectedAction>;
 } & S;
 
 export default function createPropsPicker<S, P = DefaultPickedProps<S>>(
-  propsPicker: PropsPicker<P, Partial<P>> = defaultPropsPicker,
+  propsPicker: PropsPicker<P, S, Partial<P>> = defaultPropsPicker,
   reduxInfo: CurtainReduxInfo<S>,
   mutableObj: CurtainMutableObj
 ) {
@@ -25,6 +48,10 @@ export default function createPropsPicker<S, P = DefaultPickedProps<S>>(
   let latestProps: Partial<P> | Partial<DefaultPickedProps<S>>;
   let stateHandler = {
     get: function(target: { state: any }, name: string) {
+      let levelNum = getRelativeLevel(name);
+      if (levelNum != null) {
+        return getLevelState(target.state, sceneReducerKey, levelNum);
+      }
       let dictItem = arenaReducerDict[name];
       if (dictItem == null) return null;
       return target.state[dictItem.reducerKey];
@@ -32,14 +59,18 @@ export default function createPropsPicker<S, P = DefaultPickedProps<S>>(
   };
   let actionsHandler = {
     get: function(target: { state: any }, name: string) {
+      let levelNum = getRelativeLevel(name);
+      if (levelNum != null) {
+        return getLevelState(target.state, sceneReducerKey, levelNum);
+      }
       let dictItem = arenaReducerDict[name];
       if (dictItem == null) return null;
       return dictItem.actions;
     }
   };
   let stateObj = { state: null };
-  let stateDict = new Proxy(stateObj, stateHandler);
-  let actionsDict = new Proxy(stateObj, actionsHandler);
+  let stateDict: StateDict<S> = new Proxy(stateObj, stateHandler) as any;
+  let actionsDict: ActionsDict = new Proxy(stateObj, actionsHandler);
   return (state: any) => {
     stateObj.state = state;
     if (