+++ src/hocs/ArenaScene/ArenaScene.tsx
@@ -1,7 +1,7 @@
 import * as React from "react";
 import * as PropTypes from "prop-types";
 import { ActionCreatorsMapObject, AnyAction } from "redux";
-import { EnhancedStore, ReducerDict } from "../../core";
+import { EnhancedStore, ReducerDict, SceneBundle } from "../../core";
 import ActionTypes from "../../core/ActionTypes";
 import { createCurtainReducer } from "../../core/reducers";
 import {